Rename ui/other -> ui/keyboard
