+++ src/bokeh/models/ui/keyboard.pyi
